Update Gradle build definition
This commit updates Gradle to the most recent version, 6.7.1, and
adjusts the build.gradle to take advantage of some of Gradle's new
features. The main new feature is using the new built-in methods for
generating javadoc and sources jar, rather than needing to define tasks
manually.

This commit also adjusts the build script to apply the `java-library`
plugin rather than the `java` plugin. This allows us to declare a
dependency using the `api` configuration which causes the generated pom
to list that dependency transitively.